moving viewModels, states, events from a VM package to a previous package. in addition, moving views from the list package or the rope package to the root "jumps" package
